Extract color-related code into module
